Improve entity damage with descriptive return type
As the vanilla source code never had any need for descriptive and
non-binary return values for the Entity#damageEntity method upstream
simply returned false from this method if the entity could have been
damaged but plugins prevent it.

This meant that outside of the method itself no code was capable of
identifying whether or not the entity damage event did not occure due to
vanilla mechanics or plugin interference.

The problem became apperant due to the Explosion.java code skipping
knockback for all entities that were not damaged, whether or not that
was due to plugin influence.
While vanilla would apply knockback to any entity caught in the
explosion, ignoring whether or not the entity actually takes damage,
spigot cannot do so as that would knockback entities for which the
damage event was cancelled.

This commit introduces the Entity#applyDamage method which
returns an enum constant of DAMAGED (which replaces the vanilla true),
NOT_DAMAGED (which replaces the vanilla false) and
BUKKIT_EVENT_CANCELLED which is returnd if the damage failed due to an
event.
The Explosion.java code now also uses this return value to apply
knockback to all entities for which the damage event was not cancelled.