Resurrect paper (script)
Bring back our beloved build tool, better than ever!

This is a rewrite of the original script, while trying to be as
compatible as possible. There's a few changes though:
1. Patching and path rebuilding commands accept argument to only patch
   api or server.
2. testserver command no longer setups whole server with plugins,
   instead uses runserver config. It also accept specifying whether to
   run dev server (default), shadow or reobfuscated one.

This script now has all the logic in functions, which shifts the burden
of unregistering variables onto the shell - we only need to remember to
cleanup helper functions. Allows for more often use of meaningful
variables, which should hopefully make it easier to read.

I tried to make it as compatible and portable as possible, assuming
basic BASH and ZSH features, but it's very likely that I missed
something.