[Experimentation] Use Entity#makeBoundingBox
The initial patch to prevent bounding box desync used the
EntityDimension to create the bounding box if neccesary. This fails to
respect the ability of entities overwriting their Entity#makeBoundingBox
method (e.g. what shulkers do).
This commit moves the bounding box recalculation after the position
update but prior to any callbacks to the level.