Skip loading chunk in check if chunk is loaded
I've looked at all the uses of this method, none of them expect the
method to actually load the chunk off disk if it's not loaded. This is a
solid improvement as it reduces chunk loads from entities, xray, and
even fixes the "seed-based-feature-search-loads-chunks" configuration
option which was broken before.