Add deepslate as xray engine 2 replacement default
With the addition of deepslate in 1.17, the anti xray configuration was
updated to include the deepslate variants of the ores in regards to
hiding them.
During this change, the fact that most surrounding blocks are deepslate
was overlooked and hence anti-xray engine 2 did not have any blocks to
obfuscate the ores with.

This commit fixes that by adding the deepslate block to the default
values on config generation.

Notably, this will not forceably add the deepslate default to existing
configs, hence old config files will need to be manually edited.

Resolves: #6350