Prevent duplication on overfilled bundle addition
Bundles in minecraft are, at least in version 1.17.1, limited to a
maximum weight of 64, which in most cases correlates to the amount of
items inside the bundle.
If a bundle is overfilled, e.g. was spawned in with enough items to
break the maximum weight of 64, adding more items into the bundle will
cause the bundle to instead add an air itemstack with the inverted
amount of the added item, resulting in a duplication of the original
item.

While mojang already cancels additions to the bundle if the bundle has 0
weight left, they do not cancel additions if the bundle has a negative
amount of weight left.

This commit changes the bundle to now prevent additions to the bundle if
the weight left to be filled in the bundle is zero or less, effectively
preventing the duplication.

Resolves: #6550