Separate the despawn ranges from the squared range
The patch "Optimise player lookups" uses the hard despawn distance when looking
for players. However, the method expects a normal range instead of the squared
range given. This change (goofy as it looks, since it doesn't modify the
problematic patch) ensures that the squared range is used for squared checks,
and the normal ranges are still available for non-squared checks.