Add/Remove multiple entries to Scoreboard Team
It is a common pattern to fill a team up with several entries. Unfortunately, each single entry addition sends a ScoreboardTeam packet out which can be unnecessarily chatty. This patch just adds the option to set a collection of entries on a scoreboard team which can get bundled into a single packet.

Likely not a significant optimization, but hey, every little bit helps.