Fix nullability on Block#breakNaturally
The implementation of this method allows for null tools. The use of a null tool would be useful if the block should be dropped always. E.g. bedrock can't be broken with any tool, but it will be dropped if you specify a null tool.

Additionally, the original breakNaturally method also allows for null tools and just calls the added method from paper.