Expose the potential player cause of a lightning
As the internal lightning bolt representation already stores the player
instance that may have caused the lightning in the first place, through
tridents, this commit expands the existing additions to the lightning
api to expose the previously mentioned player instance that caused the
lightning.

Additionally, a setter for the player cause has been added to the
lightning strike entity to mutation of the previously mentioned player
reference.