Correctly inflate villager activation bounding box
Prior to this commit, the activation bounding box used for villagers was
inflated using the 'entity-activation-range.villagers' for the x axis
and 'entity-activation-range.water' for the z axis, causing villagers to
use the water mobs activation range on the z axis.

This commit now properly uses the 'entity-activation-range.villagers'
value for both the x and z axis inflation for the villagers activation
range.

Resolves: #6797