Fix riding distance statistics (#7021/SPIGOT-6475)
This PR ports a slightly simplified version of https://hub.spigotmc.org/stash/projects/SPIGOT/repos/craftbukkit/pull-requests/900 to 1.18. As described in the Spigot PR, Spigot synchronizes the position of the vehicle and the player. However, these locations may not always match up. Later down the line, the player position gets set to the correct value, and the player gets the difference added to the statistic.