Makes it possible to change the xPush and zPush of powered minecarts
Adds a method for getting the push of a powered minecart as a two-dimensional vector where the first coordinate is xPush and the second is zPush.
Adds a method for setting the push of a powered minecart that expects a two-dimensional vector where the first coordinate is xPush and the second is zPush.