Properly create profiles with custom name/uuid
Ensure that the server properly creates custom player profiles if either
the uuid or name match an online player. Previously the server would
directly shortcut to the players profile if either their uuid or name
matched, ignoring differences in the respective alternative identifier
like a custom name or a different uuid.

To keep the previous functionality, the method does populate the profile
with the online players properties if either the uuid or name match.