Grant temporary immunity from EAR to moving entities
Moving entities behave erratically when EAR tick decimation activates. This
breaks e.g. ice based sorters in the Nether. To counter this a test based
on the test-for-movement in ItemEntity.tick() is added to to the
ActivationRange, granting a temporary immunity. Once the momentum is lost
the item again goes inactive.