SPIGOT-6949: Configuration sections that are nested within Maps or Lists are not properly serialized.
This broke with the configuration changes in
ed8a152b3ae55c5f2b783729ed284d14010df169.

This commit reverts one of the changes of this other commit so that
nested configuration sections are serialized as Maps again.

Although the types of these nested configuration sections are not
preserved when reloading a configuration (they turn into Maps), their
contents should at least be preserved, as it has been the case in
earlier Bukkit versions.

By: blablubbabc <lukas@wirsindwir.de>