Revert to old createProfile(UUID, String) logic
The recent change to the createProfile method set, which allowed plugins
to properly create profiles with differing names/uuids for players
currently on the server broke an implicit method contract for plugins
providing the lowercased version of a player name to the method.

Previously, the returned profile would simply be pulled from an online
player with the same name (when compared case-insensitively), however
the previously mentioned fix to the creation logic currently returns
a profile with the lowercased name instead.

To not break plugins that relied on this specific internal logic, the
ability to create pre-filled profiles with different names or uuids for
players that are currently online is moved by this commit to a new
method called createProfileExact, which simply mirrors the current
createProfile logic.

The createProfile logic is hence returned to its old functionality, once
again allowing plugins to grab an online players profile from their
lowercased name.