Call EntityDropItemEvent for goats dropping horns
Calls the EntityDropItemEvent when a goat drops one of its goat horns.

As with other calls of the EntityDropItemEvent, the event is published
after the modification of the entity and cancellation of it does not
restore the entities state. In this case, the goat would still loose
their horn however the item would not be dropped (See snowmen dropping
their pumpkin).