Re-add watchdog dumps
Also make the final watchdog dump perform a chunk debug if
we were waiting on chunks

In order to ensure the watchdog thread cannot stall on
lock acquisition, the scheduling and tick locks have been
changed to ReentrantLock so that we can use tryLock with
a time out.

Even if the watchdog thread fails to acquire the locks, it will
still try to perform a dump.