Only load in entity chunk for FULL chunk status progression
Before, EntityLookup would load in the entity check if it did
not exist. However, this was unneccessary. Instead, we just create
a transient entity chunk now.

We also want to restrain the EntitiesLoadEvent to the chunk full
status, rather than having it possibly be called for
things such as Entity#move, addEntity, etc, as we know that
the chunk full status is a safe place to call it.