Respect accessible visibility status
The new chunk system made any entity accessible whether or not
the status was accessible to ensure that when a chunk was
eventually loaded to plugins, the entities would be visible.
However, this caused issues with the dynamic game event listener,
as it has no logic to properly deal with unloaded chunks in the
move() method, as it expects the accessible hooks to call
add/remove while the chunk is still loaded.

Fortunately, these changes still ensure that when the chunk
load event is called, that the entities will be accessible.