Restore no-crash behaviour when read-only config file(s)
Prior to migrating handling configuration with Configurate,
paper did not crash when it found it couldn't overwrite the
existing paper.yml configuration file (but left a stack trace).

After the migration, paper crashes when it is unable to overwrite
any of paper-global.yml, paper-world-defaults.yml or a
world-specific config file.

This aims at returning to the previous behaviour of not
crashing when paper is unable to overwrite a configuration
file.

Signed-off-by: Antoine Mazeas <antoine@karthanis.net>