Adjust handling of mustNotSave and properly handle isUnsaved
The new isUnsaved handling also ensures we do not save
unloading chunks even if they do not need to be saved,
as was the behavior of the previous system.

The new handling of mustNotSave is to allow the entity chunk
to save, but not the poi chunk or the level chunk. While
this behavior is ridiculous, it was also (minus the poi part)
the behavior of the old chunk system which nobody complained about,
so I'm not going to bother to fix it.