Properly close section storage managed files
While nuking the mojang IO worker in the async chunk transition, the
SectionStorage's close method was missing the super call to its parent
RegionFileStorage which in turn is responsible for closing the region
files in use by the section storage. This would manifest in opened files
for an unloaded world.

This commit adapts the SectionStorage#close method to properly call its
parent's close method, closing any open files.