Properly fail on errors while reading update dir
Previously PluginInitializerManager used YamlConfiguration#load to load
the bukkit.yml file on the server which quietly fails if the file does
not exist yet.

If however an IOException that is not a FileNotFoundException is thrown
or the config file is invalid, the logic submits an error log to
Bukkit#getLogger(), which is not initialised early enough for this logic
to pick it up.

To prevent this, PluginInitializerManager was moved to use
PaperConfigurations's #loadLegacyConfigFile(File) method, which properly
throws any exception up the stack.