Resend more data on profile change
Currently breaking are potion effects and the XP bar, both are now being resent to make the client know about them again.
Because the #teleport method also updates abilities and scaled health, I'm updating those as well just in case.
It may be worth looking into other things that need to be resent in the future.