Feature: FOV Scale
Modifies the player's fov by a factor
The multiplier is affected by the client Accessibility option 'FOV
Effects'

I'm not sure about the approach of exposing a variable in
ClientboundPlayerAbilitiesPacket and then mutating it. Alternatively
the onUpdateAbilities method could clone Abilities before constructing
the packet. Let me know if that approach is preferred