Fix sendSignChange0's usage of SignText
Minecraft 1.20 introduced the SignText class that holds onto the text of
a single side of a minecraft sign.
This type is immutable and any methods on it produce an updated, new
instance of the class rather than mutating the called upon instance.

Spigot does not respect this change as it does not re-assign the
constructed sign text to the instance nor does it feed back said
instance to the virtual sign block entity.