Properly grow plants with growth modifier
Spigots implementation of the growth modifier boils down to a plain
chance for a plant to either grow immediately or to grow a single age
value towards its fully grown age.

This behaviour is specifically troubling for plants like the cactus or
the sugar cane, as they change an unrelated block when growing up (the
free air block above their tip). This blocks validity is checked during
neighbour updated which are triggered by resetting the previous tip of
the plant back to [age=0].

With spigots modifier implementation, blocks at the age of zero are able
to grow a new block above them however, causing the block change of the
previous tip to change a block of age zero to a block of age zero, which
does **not** cause a neighbour update as nothing actually changed.
This means that the about block is not updated and hence may remain in
an otherwise invalid state. (E.g. a cactus block next to a solid block).

This commit rewrites the existing spigot growth modifier implementation
specifically for the plants that grow new blocks like the cactus or the
sugar can block.

The modifier provided for the plants is separated into two distinct
values. One being the remains after a floored division by 100, e.g. the
original modifier without its 10^0 and 10^1 digits, the other the
modifier modulo 100.
The modifier 475 would, for example, be split into a base growth of
4 and a percentage chance of an additional age progression of 75%.
This way, on average, the plant would grow 4.75 times as fast as a
normal plant, which would only grow a single age per random tick.