Fix delay player chunk unloads
It was not functional as the check for a ticket level change before adding the delay ticket would never succeed. This is because the player chunk loader adds an unknown ticket at the same level before removing the player ticket (with addAndRemove).

This fixes it by moving the delay ticket add to the player chunk loader's addAndRemove call.