Call LivingEntity#onItemPickup before mutation
The existing EntityPickupItemEvent fixes patch moves the call to
LivingEntity#onItemPickup for piglins after the respective
EntityPickupItemEvent calls, which is correct.
However the patch moved the call so far down the line that the existing
logic already mutated the picked up item entity, leading to faulty state
being passed to the onItemPickup method.

To prevent logic in LivingEntity#onItemPickup to read from an ItemEntity
that was already mutated, this commit moves the calls prior to the two
respective mutations (either gold_nugget or rest).

This was chosen above taking a copy of the original item and restoring
state later on to avoid a full item stack clone.