Add new FluidState API functionality
This commit adds a new FluidState API to the system. It introduces various methods to get information from fluid states in the game, such as checking if a block is a fluid source, getting the amount of fluid, type of the fluid (water or lava), and its flow direction. It helps to enhance flexibility when managing fluid dynamics in the game. Alongside new API, it also modifies the existing `RegionAccessor` class to incorporate the functionalities of the FluidState API.