Fix CraftStructureTransformer from deleting generated entities
CraftLimitedRegion loads and stores entities in the protochunk and doesn't
parse the entities from NBT each time. This *requires* that a new one
instance of CraftLimitedRegion is created each time before passing it
to an api consumer and then saving/clearing it afterwards.

CraftStructureTransformer was creating a single instance for an entire structure place which
was causing entities to be deleted so no structures would spawn entities.