Fix Mob Effect Visibility/Attributes syncing
Currently, Minecraft waits a tick before syncing effect visiblity/glowing.
This is noticeable when adding entities into the world with effects.
So, this main logic moves the dirty effect logic into the actual updating logic.
This causes effect visibility to be updated correctly.

Additionally, mob effects are not correctly added when read from NBT due to their attributes not being applied.
This is a problem for entities that have effects already added before being ticked for the first time.
We also manually update effect visibility, causing glowing and invisibility to be correctly applied as well.